:recycle: Rename components based on PR feedback
